Revert "temp"
This reverts commit 7e2d318d815bac0346295a20085e3a69d98085d9.

+++ src/playground_base/playground_base/base/topo/param/FBDiscreteParam.hpp
@@ -12,15 +12,15 @@ struct FBDiscreteParam
 
   int ValueCount() const { return valueCount; }
 
-  double PlainToNormalized(int plain) const;
-  int NormalizedToPlain(double normalized) const;
+  float PlainToNormalized(int plain) const;
+  int NormalizedToPlain(float normalized) const;
 
   std::string PlainToText(int plain) const;
   std::optional<int> TextToPlain(std::string const& text) const;
 };
 
 inline int
-FBDiscreteParam::NormalizedToPlain(double normalized) const
+FBDiscreteParam::NormalizedToPlain(float normalized) const
 {
   return std::clamp((int)(normalized * valueCount), 0, valueCount - 1);
 }
