+++ src/playground_base/playground_base/base/topo/param/FBLinearParam.hpp
@@ -8,23 +8,23 @@
 
 struct FBLinearParam
 {
-  double min = 0.0;
-  double max = 1.0;
-  double displayMultiplier = 1.0;
+  float min = 0.0f;
+  float max = 1.0f;
+  float displayMultiplier = 1.0f;
 
   int ValueCount() const { return 0; }
 
-  std::string PlainToText(double plain) const;
-  std::optional<double> TextToPlain(std::string const& text) const;
+  std::string PlainToText(float plain) const;
+  std::optional<float> TextToPlain(std::string const& text) const;
 
-  double PlainToNormalized(double plain) const;
-  double NormalizedToPlain(double normalized) const;
+  float PlainToNormalized(float plain) const;
+  float NormalizedToPlain(float normalized) const;
   FBFloatVector NormalizedToPlain(FBFloatVector normalized) const;
   FBDoubleVector NormalizedToPlain(FBDoubleVector normalized) const;
 };
 
-inline double
-FBLinearParam::NormalizedToPlain(double normalized) const
+inline float
+FBLinearParam::NormalizedToPlain(float normalized) const
 {
   return min + (max - min) * normalized;
 }
